Major revision of page cycler UI.  This revision does several things:
In this CL:

1. Reforms the JS UI to look more like Chrome|Settings, including look/feel and use of overlays for progress and error reporting.

2. Redesigns the UI to support maintaining on the extension side a record of captures, including their URLs and the cache state resulting from their initial capture run.  Allows the user to add new captures to the record, delete unwanted ones, and run any he'd like.

Later CLs:

3. Actually implements the underlying FileSystem-using code to do this, as opposed to current mockup.

4. Adds C++-side code to allow C++ directories to be copied to and from FileSystem sandbox directories.

5. Adds means of interrupting ongoing sub-browser sessions so that captures and replays may be cancelled from UI.

6. Improves stats reporting, including nice-looking charts instead of present rather nasty line-by-line text dump of runtimes per URL.


Snapshots:
Capture tab: http://imgur.com/Jb3UG
Message overlay: http://imgur.com/afFVL
Playback tab: http://imgur.com/PUGhx
Playback w/o chosen capture: http://imgur.com/biKK1
Playback w/ no captures available: http://imgur.com/Zv4Bp

Aaron's initial mockup/spec:
https://moqups.com/aboodman/wDjKAEXR/p:aed59f081

Somewhat dated original design doc for the first version:
https://docs.google.com/a/chromium.org/document/d/1HyiPO4DBfCzLtVKOMx6rGGa7Vw_GzNZX_FObc-mWiZo/edit?hl=en_US&sai=true

Review URL: https://chromiumcodereview.appspot.com/10832191

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@162059 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 0c2286e417d1427f48a1296460b10c7e442a2f19

+++ chrome/browser/extensions/api/record/record_api.h
@@ -156,7 +156,7 @@ class ReplayURLsFunction : public RunPageCyclerFunction {
   base::Time timer_;
 
   // These two data are additional information returned to caller.
-  int run_time_ms_;
+  double run_time_ms_;
   std::string stats_;
 };
